pyproject.toml
[tool.poetry]
name = "students-rooms"
version = "0.1.0"
description = ""
authors = ["Dmitry-BigData <papko.employment@gmail.com>"]
readme = "README.md"
packages = [{include = "students_rooms"}]

[tool.poetry.dependencies]
python = "^3.10"
psycopg2 = "^2.9.6"
pandas = "^2.0.1"
python-dotenv = "^1.0.0"
sqlalchemy = "^2.0.13"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
